requirements.txt
anyjson==0.3.3
cached-property==1.4.2
certifi==2018.1.18
chardet==3.0.4
colored==1.3.5
configparser==3.5.0
coverage==4.5.1
coveralls==1.3.0
docopt==0.6.2
Faker==0.8.12
frozendict==1.2
httplib2==0.11.3
idna==2.6
mysql-connector==2.1.6
nose==1.3.7
PyMySQL==0.8.0
python-auspost-pac==1.0.2
python-dateutil==2.7.2
requests==2.18.4
six==1.11.0
text-unidecode==1.2
urllib3==1.22
